Skip running collector tests when in CI (for now), because of the connectivity issues
